pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "talent-scouting-intelligence"
version = "0.1.0"
description = "Early momentum detection for music scouting with legal low-cost data paths"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Codex"}]
dependencies = [
  "PyYAML>=6.0",
  "matplotlib>=3.7",
  "streamlit>=1.32",
  "pandas>=2.0"
]

[project.scripts]
tsi = "talent_scouting_intel.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
